Toss Random
This is first step for random Cricket. It includes 8 teams of IPL which has squad of 2020. In this you will need to give input of team 1 and team 2. Remember you can only use short form of these teams to give input. As:-
CSK
MI
KXIP
RR
SRH
DC
RCB
KKR
this will tell you who throw toss (which always team 1) and what team 2 sayed. Who won the toss and what he selected. This is fully random process. Then It will give you random Playing XI(11) for both teams. Which you can accept or Run again. The only problem in the program is in real there are only 4 foreginers allowed in one team a time. But this will give as many as it wants. If you know solution please tell me